pll.c: 6: op: MISRA.ETYPE.ASSIGN.2012
Fixed MISRA 10.3 violation by changing literal suffix from UL to ULL to
ensure compatible essential types in assignment. Changed expression from
(1UL << data->data->pllfm_bits) to (1ULL << data->data->pllfm_bits) to
match uint64_t target type. Fixed 1 violation, ID: 2072, tfa comment:
Fixing existing code may be counter-productive and introduce bugs.

+++ plat/ti/k3low/common/drivers/pm/clock/pll.c
@@ -500,7 +500,7 @@ static inline void pll_consider_fractional(struct pll_consider_data *data,
 	 *
 	 * pllm_rem * pllfm range = pllfm * input
 	 */
-	rem_target = (uint64_t) (((uint64_t) pllm_rem) * (1UL << data->data->pllfm_bits));
+	rem_target = ((uint64_t) pllm_rem) * (1ULL << data->data->pllfm_bits);
 
 	/* Start at the lowest and walk it up to the highest */
 	pllfm_input = ((uint64_t) lowest_pllfm) * data->input;
